requirements.txt
autopep8==1.4.3
Django==2.2.1
django-cors-headers==2.5.3
django-rest-auth==0.9.5
djangorestframework==3.9.3
pycodestyle==2.4.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
virtualenv==16.5.0
